backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.1
pydantic-settings==2.6.1
sqlmodel==0.0.22
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
qdrant-client==1.12.1
openai==1.55.0
python-multipart==0.0.20
better-exceptions==0.3.4
python-dotenv==1.0.1
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==43.0.3
requests==2.32.3
httpx==0.27.2
alembic==1.14.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0